--ns-cert-type depreciated
As of OpenSSL v1.1, the nsCertType extension in X.509 certificates are no longer supported. ... The replacement option, ---remote-cert-tls is a macro which sets the --remote-cert-ku and --remote-cert-eku to appropriate values, depending on it is wanted to check if the remote provided certificate is a server or client certificate. As the extended key usage extension is far more commonly used today, this is effectively the equivalent of --ns-cert-type.
https://community.openvpn.net/openvpn/wiki/DeprecatedOptions#a--ns-cert-type